Change the mixin from Fabric to common folder.
