+++ common/src/main/resources/flan.mixins.json
@@ -19,6 +19,7 @@
     "EntityMixin",
     "FireBlockMixin",
     "FluidMixin",
+    "GiveCommandMixin",
     "IHungerAccessor",
     "ILecternBlockValues",
     "IPersistentProjectileVars",
