+++ fabric/src/main/resources/flan.fabric.mixins.json
@@ -7,7 +7,6 @@
     "EntityDamageMixin",
     "FabricFireMixin",
     "LevelMixin",
-	  "GiveCommandMixin",
     "LightningHitMixin",
     "PatrolSpawnerMixin",
     "PhantomSpawnerMixin",
